third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-shapes/spec-examples/shape-outside-012-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Shape from image - shape-image-threshold - 0.9 Uncaught Error: assert_approx_equals: Line 0 is positioned properly expected 189 +/- 2 but got 179
Harness: the test ran to completion.
